+++ util/src/main/java/tech/pegasys/artemis/util/backing/Utils.java
@@ -15,7 +15,7 @@ package tech.pegasys.artemis.util.backing;
 
 public class Utils {
 
-  public static int nextPowerOf2(int x) {
-    return x <= 1 ? 1 : Integer.highestOneBit(x - 1) << 1;
+  public static long nextPowerOf2(long x) {
+    return x <= 1 ? 1 : Long.highestOneBit(x - 1) << 1;
   }
 }
